gicv5_its_intr: fix argument type
Both the caller and its use want unsigned long to use that instead of
uintptr_t.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK

+++ sys/arm64/arm64/gicv5_its.c
@@ -555,7 +555,7 @@ its_device_get(device_t dev, struct its_device_list *dev_list, device_t child,
 
 
 static int
-gicv5_its_intr(void *arg, uintptr_t irq)
+gicv5_its_intr(void *arg, unsigned long irq)
 {
 	struct gicv5_its_softc *sc = arg;
 	struct gicv5_its_irqsrc *gi;
